Stars and Pause bugs
Added in the stars background, but perhaps not complete.
Also identified and implemented a fix for a bug with the pause button.
If you paused while feedback was displaying, the circle would reset
behind the screen and eventually get game over. The fix (for now) is
disabling the pause button while feedback displays (until the circle
resets).